Replace MT5Account.cs with NuGet v1.0.942 and update documentation
Core Changes:
- Replaced MT5Account.cs with decompiled version from NuGet package v1.0.942 (1729 lines)
- Fixed namespace in 74 documentation files: MetaRPC.CSharpMT5 → mt5_term_api
- Fixed namespace in Your_First_Project.ru.md (using mt5_term_api)

Code Fixes:
- Added (uint) casts in MT5Service.cs for sessionIndex parameter (lines 785, 800)
- Fixed examples: Service (removed OrderCalcProfitAsync), Monitor (added 5 sec delay), Scalper (increased SL/TP distance to 30 points)
- Excluded ReflectTool and ReflectMT5Account.cs from compilation

API Documentation:
- Removed AttachByInstanceId from MT5Account.API.md (method not in NuGet v1.0.942)
- Fixed sessionIndex parameter type: int → uint in 4 methods (SymbolInfoSession*)
- Updated method count: 72 → 71

All 13 examples tested and working

+++ MT5Service.cs
@@ -782,7 +782,7 @@ namespace MetaRPC.CSharpMT5
             int sessionIndex,
             DateTime? deadline = null,
             CancellationToken cancellationToken = default)
-            => _acc.SymbolInfoSessionQuoteAsync(symbol, dayOfWeek, sessionIndex, deadline, cancellationToken);
+            => _acc.SymbolInfoSessionQuoteAsync(symbol, dayOfWeek, (uint)sessionIndex, deadline, cancellationToken);
 
         /// <summary>Gets trading session trade information for the specified symbol and day.</summary>
         /// <param name="symbol">Target symbol name (e.g., "EURUSD").</param>
@@ -797,7 +797,7 @@ namespace MetaRPC.CSharpMT5
             int sessionIndex,
             DateTime? deadline = null,
             CancellationToken cancellationToken = default)
-            => _acc.SymbolInfoSessionTradeAsync(symbol, dayOfWeek, sessionIndex, deadline, cancellationToken);
+            => _acc.SymbolInfoSessionTradeAsync(symbol, dayOfWeek, (uint)sessionIndex, deadline, cancellationToken);
 
         /// <summary>Calculates tick value and size for one or multiple symbols.</summary>
         /// <param name="symbols">Collection of symbol names (e.g., "EURUSD", "GBPUSD").</param>
